requirements.txt
@"
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
qdrant-client==1.7.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-classic==0.0.1
openai-whisper==20231117
edge-tts==6.1.9
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
psutil==5.9.5
torch==2.1.0
numpy==1.24.3
pyttsx3==2.90
pypiwin32==223
"@ | Out-File -FilePath requirements.txt -Encoding utf8